+++ src/libranet_logging/cli.py
@@ -1,33 +1,28 @@
 # pylint: disable=import-outside-toplevel
 """libranet_logging.cli"""
 import logging
-import os
 
-import click
 import logging_tree
 
-from libranet_logging.logconfig import get_default_logging_yml, initialize
+# @click.command()
+# @click.argument("path", envvar="PYTHON_LOG_CONFIG", required=False, type=click.Path(exists=False))
+# def print_logging_tree(path) -> None:
+#     """Initializes the logging and print the configured logging-tree."""
+#     os.environ["PYTHON_ENABLE_LOGGING_TREE"] = "True"
 
+#     if not path:
+#         click.secho(
+#             "No path provided to a logging.yml and the env-var PYTHON_LOG_CONFIG is not set.",
+#             fg="yellow",
+#             bold=True,
+#         )
+#         click.secho(
+#             f"Using the packaged default from {get_default_logging_yml()}",
+#             fg="yellow",
+#             bold=True,
+#         )
 
-@click.command()
-@click.argument("path", envvar="PYTHON_LOG_CONFIG", required=False, type=click.Path(exists=False))
-def print_logging_tree(path) -> None:
-    """Initializes the logging and print the configured logging-tree."""
-    os.environ["PYTHON_ENABLE_LOGGING_TREE"] = "True"
-
-    if not path:
-        click.secho(
-            "No path provided to a logging.yml and the env-var PYTHON_LOG_CONFIG is not set.",
-            fg="yellow",
-            bold=True,
-        )
-        click.secho(
-            f"Using the packaged default from {get_default_logging_yml()}",
-            fg="yellow",
-            bold=True,
-        )
-
-    initialize(use_print=True)
+#     initialize(use_print=True)
 
 
 def print_tree() -> None: